+++ packages/dnr-rulesets/CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+<!-- 4.2.0-beta.2 is the same as 4.2.0-beta.1, just a new build with latest filters -->
+
 ## [4.2.0-beta.1] - 2026-04-17
 
 ### Added